Consoildate demo and tutorial stuff under demo. Fixes #1169
