+++ ziti/cmd/demo/plain_echo_client.go
@@ -14,7 +14,7 @@
 	limitations under the License.
 */
 
-package tutorial
+package demo
 
 import (
 	"fmt"
