+++ ziti/cmd/demo/setup-multi-tunneler-hosted-service.go
@@ -23,7 +23,6 @@ import (
 	"github.com/openziti/ziti/ziti/cmd/api"
 	"github.com/openziti/ziti/ziti/cmd/common"
 	cmdhelper "github.com/openziti/ziti/ziti/cmd/helpers"
-	"github.com/openziti/ziti/ziti/cmd/tutorial"
 	"github.com/spf13/cobra"
 	"time"
 )
@@ -33,7 +32,7 @@ var multiTunnelerHostedScriptSource []byte
 
 type multiTunnelerHosted struct {
 	api.Options
-	tutorial.TutorialOptions
+	TutorialOptions
 	interactive bool
 }
