+++ ziti/cmd/demo/ziti_echo_server_cmd.go
@@ -14,39 +14,41 @@
 	limitations under the License.
 */
 
-package tutorial
+package demo
 
 import (
 	"github.com/openziti/ziti/ziti/cmd/common"
 	cmdhelper "github.com/openziti/ziti/ziti/cmd/helpers"
 	"github.com/spf13/cobra"
-	"strings"
+	"time"
 )
 
 // Options are common options for edge controller commands
-type plainEchoClientOptions struct {
+type zitiEchoServerOptions struct {
 	common.CommonOptions
-	port uint16
-	host string
+	identity string
 }
 
-func (self *plainEchoClientOptions) run() error {
-	echoClient := &plainEchoClient{
-		host: self.host,
-		port: self.port,
+func (self *zitiEchoServerOptions) run() error {
+	echoServer := &zitiEchoServer{
+		identityJson: self.identity,
 	}
-	return echoClient.echo(strings.Join(self.Args, " "))
+	if err := echoServer.run(); err != nil {
+		return err
+	}
+	time.Sleep(time.Hour * 24 * 365 * 100)
+	return nil
 }
 
-func newPlainEchoClientCmd(p common.OptionsProvider) *cobra.Command {
-	options := &plainEchoClientOptions{
+func newZitiEchoServerCmd(p common.OptionsProvider) *cobra.Command {
+	options := &zitiEchoServerOptions{
 		CommonOptions: p(),
 	}
 
 	cmd := &cobra.Command{
-		Use:   "plain-echo-client strings to echo",
-		Short: "Runs a simple http echo client",
-		Args:  cobra.MinimumNArgs(1),
+		Use:   "ziti-echo-server",
+		Short: "Runs a ziti-based http echo service",
+		Args:  cobra.ExactArgs(0),
 		Run: func(cmd *cobra.Command, args []string) {
 			options.Cmd = cmd
 			options.Args = args
@@ -58,8 +60,7 @@ func newPlainEchoClientCmd(p common.OptionsProvider) *cobra.Command {
 
 	// allow interspersing positional args and flags
 	cmd.Flags().SetInterspersed(true)
-	cmd.Flags().Uint16Var(&options.port, "port", 80, "Specify the port to dial on")
-	cmd.Flags().StringVar(&options.host, "host", "localhost", "Specify the host to connect to")
+	cmd.Flags().StringVar(&options.identity, "identity", "echo-server.json", "Specify the config file to use")
 
 	return cmd
 }
